+++ tools/importer/case-study-article.js
@@ -56,33 +56,8 @@ const generateEndColumns = (document) => {
   });
 };
 
-// Specifically for the success section
-const standardArticleInitialColumns = (document) => {
-  const columns = document.querySelector('.no-gutters.justify-content-start');
-  if (columns) {
-    const cells = [['Columns']];
-    const arr = [];
-
-    const imgSrc = columns.querySelector('img')?.src;
-    if (imgSrc.indexOf('ribbon') !== -1) {
-      [...columns.children].forEach((child, index) => {
-        if (index === 0) {
-          arr.push(':ribbon:');
-        } else {
-          arr.push(child.innerHTML);
-        }
-      });
-
-      cells.push(arr);
-      const table = WebImporter.DOMUtils.createTable(cells, document);
-      columns.replaceWith(table);
-    }
-  }
-};
-
 export {
   threeColumnsArticleXS,
-  standardArticleInitialColumns,
   generalColumns,
   generateEndColumns,
 };
